Change MultiTag to accept Object instead of Integer
Change MultiTag to allow Object in order to store values
other than Integer.

Bug: 148036411
Test: MutableTagBundleTest / SessionConfigTest

+++ camera/camera-camera2/src/main/java/androidx/camera/camera2/internal/Camera2CameraControlImpl.java
@@ -450,7 +450,7 @@ public class Camera2CameraControlImpl implements CameraControlInternal {
         mSessionConfigBuilder.setImplementationOptions(getSessionOptions());
         Object tag = mCamera2CameraControl.getCamera2ImplConfig().getCaptureRequestTag(null);
         if (tag != null && tag instanceof Integer) {
-            mSessionConfigBuilder.addTag(Camera2CameraControl.TAG_KEY, (Integer) tag);
+            mSessionConfigBuilder.addTag(Camera2CameraControl.TAG_KEY, tag);
         }
         return mSessionConfigBuilder.build();
     }
